requirements.txt
# Leon Trading Bot - Зависимости
# Основные зависимости
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# Визуализация
flask>=3.0.0
plotly>=6.0.0
rich>=13.0.0

# Обмен данными
websocket-client>=1.7.0
ccxt>=4.0.0

# Машинное обучение
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.0.0
tqdm>=4.60.0
scipy>=1.7.0
tensorflow>=2.8.0
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0
statsmodels>=0.13.0
plotly>=5.5.0 